requirements.txt
alembic==1.4.2
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
Mako==1.1.3
MarkupSafe==1.1.1
nltk==3.5
psycopg2==2.8.5
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
redis==3.5.3
regex==2020.7.14
requests==2.24.0
rq==1.5.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.18
tqdm==4.48.0
urllib3==1.25.10
Werkzeug==1.0.1
